req/requirements.txt
joblib==1.2.0
matplotlib==3.7.1
numpy==1.24.2
pandas==1.5.3
plotly==5.13.0
PyMySQL==1.0.2
python_bcrypt==0.3.2
scikit_learn==1.2.2
scipy==1.10.1
streamlit==1.20.0
streamlit_authenticator==0.2.1
